requirements.txt
GitPython==3.1.30
pydriller==2.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
radon==6.0.1
matplotlib==3.7.2
plotly==5.17.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
tokenizers==0.14.1
seaborn==0.12.2
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
tqdm==4.66.1
joblib==1.3.2